Removed scene dependency from CameraPanelInspector
