+++ src/js/shot-generator-new/components/ItemList/index.js
@@ -7,7 +7,8 @@ const {dialog} = remote
 import {
   getSceneObjects,
   getSelections,
-  getActiveCamera, selectObject, deleteObjects, updateObject
+  getActiveCamera, selectObject, deleteObjects, updateObject,
+  setActiveCamera
 } from './../../../shared/reducers/shot-generator'
 
 import deepEqualSelector from './../../../utils/deepEqualSelector'
@@ -46,7 +47,7 @@ const isSelected = (id, selections, children = []) => {
   return unselectedChildren.length === 0
 }
 
-const ItemList = React.memo(({sceneObjects, selections, activeCamera, selectObject, deleteObjects, updateObject}) => {
+const ItemList = React.memo(({sceneObjects, selections, activeCamera, selectObject, deleteObjects, updateObject, setActiveCamera}) => {
   const listRef = useRef(null)
   
   console.log('ItemList Rerender')
@@ -70,6 +71,9 @@ const ItemList = React.memo(({sceneObjects, selections, activeCamera, selectObje
     }
     
     selectObject([...new Set(currentSelections)])
+    if(props.type === "camera") {
+      setActiveCamera(props.id)
+    }
   }, [selections])
 
   const onHideItem = useCallback((event, props) => {
@@ -174,7 +178,8 @@ const mapStateToProps = (state) => ({
 const mapDispatchToProps = {
   selectObject,
   deleteObjects,
-  updateObject
+  updateObject,
+  setActiveCamera 
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(ItemList)
